C2000 DSP28027 1. NEW: add diff_array[].

+++  sam3s-demo-source --username kaiser.ren@gmail.com/DSP2802x/v126/DSP2802x_examples/i2c_eeprom/main.c
@@ -31,7 +31,7 @@ void button_init(void);
 void main(void)
 {
 	char s1[32];
-	float floatvar1 = 23.456;
+	double floatvar1 = 23.456;
 	int intvar1 = 1234;
 
 // Step 1. Initialize System Control:
